mod: merge balancing system and helper
The difference between the two is too thing to seperate them in
different files.
